Add files via upload
We have N passengers and a tourist bus with C seats (where C < N). The behavior of the bus follows the rules:
(a) Waits to be full (until C passengers have boarded).
(b) Takes a tour of the city.
(c) Arrives back at the starting point and waits for all passengers who boarded to disembark before returning to (a).
The behavior of a passenger follows the rules:
(a') Tries to board the bus if there is still space, otherwise waits passively.
(b') Once aboard, takes the tourist tour on the bus.
(c') Disembarks from the bus and returns to (a').
Using locks, conditional variables, or semaphores, program this system where the bus and each passenger are represented by a thread. Obviously, to test your implementation, you will need to set values for C and N.
Hints: You can use a shared counter to count the number of people boarding the bus, and you need to consider who should 'wake up' the waiting entities (for example, the last passenger to board the bus will wake up the bus waiting)